chore: refactor and better docs
